// src/payload-types.ts
export interface Image {
  id: string;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "portfolios".
 */
export interface Portfolio {
  id: string;
  name: string;
  image?: (string | null) | Image;
  links?: Link[] | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "Link".
 */
export interface Link {
  text: string;
  url: string;
  type: 'internal' | 'external' | 'email' | 'phone';
  id?: string | null;
  blockName?: string | null;
  blockType: 'link_block';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "themes".
 */
export interface Theme {
  id: string;
  active?: boolean | null;
  portfolio: string | Portfolio;
  name: string;
  layout?:
    | (
        | NavbarSegmentBlock
        | HeaderSegmentBlock
        | ContactSegmentBlock
        | FeaturesSegmentBlock
        | GitHubSegmentBlock
        | FooterSegmentBlock
      )[]
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "NavbarSegmentBlock".
 */
export interface NavbarSegmentBlock {
  call_to_action?: Link[] | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'navbar_segment';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "HeaderSegmentBlock".
 */
export interface HeaderSegmentBlock {
  greeting?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  about?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  call_to_action?: Link[] | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'header_segment';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "ContactSegmentBlock".
 */
export interface ContactSegmentBlock {
  title?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'contact_segment';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "FeaturesSegmentBlock".
 */
export interface FeaturesSegmentBlock {
  title?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  features?:
    | {
        title: string;
        description: {
          root: {
            type: string;
            children: {
              type: string;
              version: number;
              [k: string]: unknown;
            }[];
            direction: ('ltr' | 'rtl') | null;
            format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
            indent: number;
            version: number;
          };
          [k: string]: unknown;
        };
        id?: string | null;
      }[]
    | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'features_segment';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "GitHubSegmentBlock".
 */
export interface GitHubSegmentBlock {
  title?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  image?: (string | null) | Image;
  call_to_action?: Link[] | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'github_segment';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "FooterSegmentBlock".
 */
export interface FooterSegmentBlock {
  text?: string | null;
  id?: string | null;
  blockName?: string | null;
  blockType: 'footer_segment';
}